+++ section/experience.tsx
@@ -1,9 +1,9 @@
 'use client'
 
 import React from 'react'
-import SectionHeading from './section-heading'
+import SectionHeading from '@/components/section-heading'
 import { useSectionInView } from '@/hook/use-section-inview';
-import CustomTimeline from './CustomTimeline';
+import CustomTimeline from '@/components/CustomTimeline';
 
 const Experience = () => {
   const { ref } = useSectionInView("Experience");
